DevOpsTcsMicro/Services/ApiManagerService.cs
using Microsoft.IdentityModel.Tokens;
using System.IdentityModel.Tokens.Jwt;
using System.Security.Claims;
using System.Text;

namespace DevOpsTcsMicro.Services
{
    public class ApiManagerService : IApiManagerService
    {
        private readonly IConfiguration _configuration;
        private readonly string ApiKey;
        private readonly string SecretKey;

        public ApiManagerService(IConfiguration configuration)
        {
            _configuration = configuration;
            ApiKey = _configuration["ApiKey"];
            SecretKey = _configuration["SecretKey"];
        }

        public bool ValidateApiKey(string apiKey)
        {
            return apiKey == ApiKey;
        }

        public string GenerateJwtToken(string recipient)
        {
            var tokenHandler = new JwtSecurityTokenHandler();
            var key = Encoding.ASCII.GetBytes(SecretKey);
            var tokenDescriptor = new SecurityTokenDescriptor
            {
                Subject = new ClaimsIdentity(new[] { new Claim("recipient", recipient) }),
                Expires = DateTime.UtcNow.AddMinutes(5),
                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
            };
            var token = tokenHandler.CreateToken(tokenDescriptor);
            return tokenHandler.WriteToken(token);
        }

        public bool ValidateJwtToken(string token)
        {
            var tokenHandler = new JwtSecurityTokenHandler();
            var key = Encoding.ASCII.GetBytes(SecretKey);
            try
            {
                tokenHandler.ValidateToken(token, new TokenValidationParameters
                {
                    ValidateIssuerSigningKey = true,
                    IssuerSigningKey = new SymmetricSecurityKey(key),
                    ValidateIssuer = false,
                    ValidateAudience = false
                }, out SecurityToken validatedToken);

                return true;
            }
            catch
            {
                return false;
            }
        }
    }
    
    
}
